fix(editor): hide new doc card when query banner is shown
